Edit default date if not argument parameters.

+++ PowerPositionReporter/Models/Trade.cs
@@ -3,6 +3,6 @@ namespace PowerPositionReporter.Models
     public class Trade
     {
         public DateTime UtcTimestamp { get; set; }
-        public double Volume { get; set; }
+        public int Volume { get; set; }
     }
 }
